Rename costBreakdownChart as costChart
