+++ src/pages/details/components/costChart/costChart.styles.ts
